requirements.txt
Django==1.7.0
djangorestframework==2.3.14
django-filter==0.7
Markdown==2.4.1
